+++ spec/lib/aixm/component/runway_spec.rb
@@ -94,6 +94,12 @@ describe AIXM::Component::Runway do
           <codePcnMaxTirePressure>W</codePcnMaxTirePressure>
           <codePcnEvalMethod>T</codePcnEvalMethod>
           <txtPcnNote>Paved shoulder on 2.5m on each side of the RWY.</txtPcnNote>
+          <valSiwlWeight>1500</valSiwlWeight>
+          <uomSiwlWeight>KG</uomSiwlWeight>
+          <valSiwlTirePressure>0.5</valSiwlTirePressure>
+          <uomSiwlTirePressure>MPA</uomSiwlTirePressure>
+          <valAuwWeight>30</valAuwWeight>
+          <uomAuwWeight>T</uomAuwWeight>
           <codeSts>CLSD</codeSts>
           <txtRmk>Markings eroded</txtRmk>
         </Rwy>
@@ -177,7 +183,7 @@ describe AIXM::Component::Runway do
     it "builds correct minimal OFMX" do
       AIXM.ofmx!
       %i(length width status remarks).each { |a| subject.send(:"#{a}=", nil) }
-      %i(composition preparation condition pcn remarks).each { |a| subject.surface.send(:"#{a}=", nil) }
+      %i(composition preparation condition pcn siwl_weight siwl_tire_pressure auw_weight remarks).each { |a| subject.surface.send(:"#{a}=", nil) }
       subject.to_xml.must_equal <<~END
         <Rwy>
           <RwyUid>