+++ spec/lib/aixm/feature/airport_spec.rb
@@ -242,6 +242,12 @@ describe AIXM::Feature::Airport do
           <codePcnMaxTirePressure>W</codePcnMaxTirePressure>
           <codePcnEvalMethod>T</codePcnEvalMethod>
           <txtPcnNote>Paved shoulder on 2.5m on each side of the RWY.</txtPcnNote>
+          <valSiwlWeight>1500</valSiwlWeight>
+          <uomSiwlWeight>KG</uomSiwlWeight>
+          <valSiwlTirePressure>0.5</valSiwlTirePressure>
+          <uomSiwlTirePressure>MPA</uomSiwlTirePressure>
+          <valAuwWeight>30</valAuwWeight>
+          <uomAuwWeight>T</uomAuwWeight>
           <codeSts>CLSD</codeSts>
           <txtRmk>Markings eroded</txtRmk>
         </Rwy>
@@ -338,6 +344,12 @@ describe AIXM::Feature::Airport do
           <codePcnMaxTirePressure>W</codePcnMaxTirePressure>
           <codePcnEvalMethod>U</codePcnEvalMethod>
           <txtPcnNote>Cracks near the center</txtPcnNote>
+          <valSiwlWeight>1500</valSiwlWeight>
+          <uomSiwlWeight>KG</uomSiwlWeight>
+          <valSiwlTirePressure>0.5</valSiwlTirePressure>
+          <uomSiwlTirePressure>MPA</uomSiwlTirePressure>
+          <valAuwWeight>8</valAuwWeight>
+          <uomAuwWeight>T</uomAuwWeight>
           <txtProfile>Northwest from RWY 12/30</txtProfile>
           <txtMarking>Dashed white lines</txtMarking>
           <codeSts>OTHER</codeSts>
@@ -373,6 +385,12 @@ describe AIXM::Feature::Airport do
           <codePcnMaxTirePressure>W</codePcnMaxTirePressure>
           <codePcnEvalMethod>U</codePcnEvalMethod>
           <txtPcnNote>Cracks near the center</txtPcnNote>
+          <valSiwlWeight>1500</valSiwlWeight>
+          <uomSiwlWeight>KG</uomSiwlWeight>
+          <valSiwlTirePressure>0.5</valSiwlTirePressure>
+          <uomSiwlTirePressure>MPA</uomSiwlTirePressure>
+          <valAuwWeight>8</valAuwWeight>
+          <uomAuwWeight>T</uomAuwWeight>
           <codeClassHel>1</codeClassHel>
           <txtMarking>Continuous white lines</txtMarking>
           <codeSts>OTHER</codeSts>